Refactor the DFSPH viscosity to its own file/struct.
