+++ src/lib.rs
@@ -43,12 +43,36 @@ extern crate nphysics2d as nphysics;
 #[cfg(all(feature = "dim3", feature = "nphysics"))]
 extern crate nphysics3d as nphysics;
 
+macro_rules! par_iter {
+    ($t: expr) => {{
+        #[cfg(not(feature = "parallel"))]
+        let it = $t.iter();
+
+        #[cfg(feature = "parallel")]
+        let it = $t.par_iter();
+        it
+    }};
+}
+
+macro_rules! par_iter_mut {
+    ($t: expr) => {{
+        #[cfg(not(feature = "parallel"))]
+        let it = $t.iter_mut();
+
+        #[cfg(feature = "parallel")]
+        let it = $t.par_iter_mut();
+        it
+    }};
+}
+
 #[cfg(feature = "nphysics")]
 pub mod coupling;
 pub mod geometry;
 pub mod kernel;
 mod liquid_world;
 pub mod object;
+#[cfg(feature = "sampling")]
+pub mod sampling;
 pub mod solver;
 mod timestep_manager;
