+++ src/solver/dfsph_solver.rs
@@ -9,100 +9,8 @@ use crate::geometry::{ContactManager, ParticlesContacts};
 use crate::kernel::{CubicSplineKernel, Kernel, Poly6Kernel, SpikyKernel};
 use crate::math::{Vector, DIM, SPATIAL_DIM};
 use crate::object::{Boundary, Fluid};
-
-macro_rules! par_iter {
-    ($t: expr) => {{
-        #[cfg(not(feature = "parallel"))]
-        let it = $t.iter();
-
-        #[cfg(feature = "parallel")]
-        let it = $t.par_iter();
-        it
-    }};
-}
-
-macro_rules! par_iter_mut {
-    ($t: expr) => {{
-        #[cfg(not(feature = "parallel"))]
-        let it = $t.iter_mut();
-
-        #[cfg(feature = "parallel")]
-        let it = $t.par_iter_mut();
-        it
-    }};
-}
-
-#[cfg(feature = "dim2")]
-type BetaMatrix<N> = na::Matrix3<N>;
-#[cfg(feature = "dim3")]
-type BetaMatrix<N> = na::Matrix6<N>;
-#[cfg(feature = "dim2")]
-type BetaGradientMatrix<N> = na::Matrix3x2<N>;
-#[cfg(feature = "dim3")]
-type BetaGradientMatrix<N> = na::Matrix6x3<N>;
-#[cfg(feature = "dim2")]
-type StrainRate<N> = na::Vector3<N>;
-#[cfg(feature = "dim3")]
-type StrainRate<N> = na::Vector6<N>;
-
-#[derive(Copy, Clone, Debug)]
-struct StrainRates<N: RealField> {
-    target: StrainRate<N>,
-    error: StrainRate<N>,
-}
-
-impl<N: RealField> StrainRates<N> {
-    fn new() -> Self {
-        Self {
-            target: StrainRate::zeros(),
-            error: StrainRate::zeros(),
-        }
-    }
-}
-
-fn compute_strain_rate<N: RealField>(gradient: &Vector<N>, v_ji: &Vector<N>) -> StrainRate<N> {
-    let _2: N = na::convert(2.0f64);
-
-    #[cfg(feature = "dim3")]
-    return StrainRate::new(
-        _2 * v_ji.x * gradient.x,
-        _2 * v_ji.y * gradient.y,
-        _2 * v_ji.z * gradient.z,
-        v_ji.x * gradient.y + v_ji.y * gradient.x,
-        v_ji.x * gradient.z + v_ji.z * gradient.x,
-        v_ji.y * gradient.z + v_ji.z * gradient.y,
-    );
-
-    #[cfg(feature = "dim2")]
-    return StrainRate::new(
-        _2 * v_ji.x * gradient.x,
-        _2 * v_ji.y * gradient.y,
-        v_ji.x * gradient.y + v_ji.y * gradient.x,
-    );
-}
-
-fn compute_gradient_matrix<N: RealField>(gradient: &Vector<N>) -> BetaGradientMatrix<N> {
-    let _2: N = na::convert(2.0f64);
-
-    #[cfg(feature = "dim2")]
-    #[rustfmt::skip]
-    return BetaGradientMatrix::new(
-        gradient.x * _2, N::zero(),
-        N::zero(), gradient.y * _2,
-        gradient.y, gradient.x,
-    );
-
-    #[cfg(feature = "dim3")]
-    #[rustfmt::skip]
-    return BetaGradientMatrix::new(
-        gradient.x * _2, N::zero(), N::zero(),
-        N::zero(), gradient.y * _2, N::zero(),
-        N::zero(), N::zero(), gradient.z * _2,
-        gradient.y, gradient.x, N::zero(),
-        gradient.z, N::zero(), gradient.x,
-        N::zero(), gradient.z, gradient.y,
-    );
-}
+use crate::solver::Akinci2013SurfaceTension;
+use crate::solver::DFSPHViscosity;
 
 /// A Position Based Fluid solver.
 pub struct DFSPHSolver<
@@ -117,12 +25,9 @@ pub struct DFSPHSolver<
     max_divergence_iter: usize,
     max_divergence_error: N,
     min_neighbors_for_divergence_solve: usize,
-    min_viscosity_iter: usize,
-    max_viscosity_iter: usize,
-    max_viscosity_error: N,
+    viscosity: DFSPHViscosity<N>,
+    surface_tension: Akinci2013SurfaceTension<N>,
     alphas: Vec<Vec<N>>,
-    betas: Vec<Vec<BetaMatrix<N>>>,
-    strain_rates: Vec<Vec<StrainRates<N>>>, // Contains (target rate, error rate)
     densities: Vec<Vec<N>>,
     predicted_densities: Vec<Vec<N>>,
     divergences: Vec<Vec<N>>,
@@ -147,13 +52,10 @@ where
             min_divergence_iter: 1,
             max_divergence_iter: 50,
             max_divergence_error: na::convert(0.1),
-            min_viscosity_iter: 1,
-            max_viscosity_iter: 50,
-            max_viscosity_error: na::convert(0.01),
             min_neighbors_for_divergence_solve: if DIM == 2 { 6 } else { 20 },
+            viscosity: DFSPHViscosity::new(),
+            surface_tension: Akinci2013SurfaceTension::new(),
             alphas: Vec::new(),
-            betas: Vec::new(),
-            strain_rates: Vec::new(),
             densities: Vec::new(),
             predicted_densities: Vec::new(),
             divergences: Vec::new(),
@@ -235,8 +137,6 @@ where
     pub fn init_with_fluids(&mut self, fluids: &[Fluid<N>]) {
         // Resize every buffer.
         self.alphas.resize(fluids.len(), Vec::new());
-        self.betas.resize(fluids.len(), Vec::new());
-        self.strain_rates.resize(fluids.len(), Vec::new());
         self.densities.resize(fluids.len(), Vec::new());
         self.predicted_densities.resize(fluids.len(), Vec::new());
         self.divergences.resize(fluids.len(), Vec::new());
@@ -245,35 +145,31 @@ where
 
         for (
             fluid,
-            (
-                alphas,
-                betas,
-                strain_rates,
-                densities,
-                predicted_densities,
-                divergences,
-                velocity_changes,
-                nonpressure_forces,
-            ),
-        ) in fluids.iter().zip(itertools::multizip((
+            alphas,
+            densities,
+            predicted_densities,
+            divergences,
+            velocity_changes,
+            nonpressure_forces,
+        ) in itertools::multizip((
+            fluids.iter(),
             self.alphas.iter_mut(),
-            self.betas.iter_mut(),
-            self.strain_rates.iter_mut(),
             self.densities.iter_mut(),
             self.predicted_densities.iter_mut(),
             self.divergences.iter_mut(),
             self.velocity_changes.iter_mut(),
             self.nonpressure_forces.iter_mut(),
-        ))) {
+        )) {
             alphas.resize(fluid.num_particles(), N::zero());
-            betas.resize(fluid.num_particles(), BetaMatrix::zeros());
-            strain_rates.resize(fluid.num_particles(), StrainRates::new());
             densities.resize(fluid.num_particles(), N::zero());
             predicted_densities.resize(fluid.num_particles(), N::zero());
             divergences.resize(fluid.num_particles(), N::zero());
             velocity_changes.resize(fluid.num_particles(), Vector::zeros());
             nonpressure_forces.resize(fluid.num_particles(), Vector::zeros());
         }
+
+        self.viscosity.init_with_fluids(fluids);
+        self.surface_tension.init_with_fluids(fluids);
     }
 
     /// Initialize this solver with the given boundaries.
@@ -746,240 +642,6 @@ where
         self.integrate_nonpressure_forces(dt, fluids);
     }
 
-    // NOTE: this actually computes beta / density_i^3
-    fn compute_betas(
-        &mut self,
-        inv_dt: N,
-        fluid_fluid_contacts: &[ParticlesContacts<N>],
-        fluid_boundary_contacts: &[ParticlesContacts<N>],
-        fluids: &[Fluid<N>],
-    ) {
-        let boundaries_volumes = &self.boundaries_volumes;
-        let _2: N = na::convert(2.0f64);
-
-        for fluid_id in 0..fluids.len() {
-            let fluid_fluid_contacts = &fluid_fluid_contacts[fluid_id];
-            let fluid_boundary_contacts = &fluid_boundary_contacts[fluid_id];
-            let betas_i = &mut self.betas[fluid_id];
-            let fluid_i = &fluids[fluid_id];
-
-            par_iter_mut!(betas_i).enumerate().for_each(|(i, beta_i)| {
-                let mut grad_sum = BetaGradientMatrix::zeros();
-                let mut squared_grad_sum = BetaMatrix::zeros();
-
-                for c in fluid_fluid_contacts.particle_contacts(i) {
-                    let mat = compute_gradient_matrix(&c.gradient);
-                    let grad_i = mat * (fluids[c.j_model].particle_mass(c.j) / _2);
-                    squared_grad_sum += grad_i * grad_i.transpose();
-                    grad_sum += grad_i;
-                }
-
-                //                for c in fluid_boundary_contacts.particle_contacts(i) {
-                //                    let mat = compute_gradient_matrix(&c.gradient);
-                //                    let grad_i = mat * (boundaries_volumes[c.j_model][c.j] * fluid_i.density0 / _2);
-                //                    squared_grad_sum += grad_i * grad_i.transpose();
-                //                    grad_sum += grad_i;
-                //                }
-
-                let mut denominator = squared_grad_sum + grad_sum * grad_sum.transpose();
-
-                // Preconditionner.
-                let mut inv_diag = denominator.diagonal();
-                inv_diag.apply(|n| {
-                    if n.abs() < na::convert(1.0e-6) {
-                        N::one()
-                    } else {
-                        N::one() / n
-                    }
-                });
-
-                for i in 0..SPATIAL_DIM {
-                    denominator.column_mut(i).component_mul_mut(&inv_diag);
-                }
-
-                *beta_i = denominator
-                    .try_inverse()
-                    .unwrap_or_else(|| BetaMatrix::zeros());
-
-                for i in 0..SPATIAL_DIM {
-                    let mut col = beta_i.column_mut(i);
-                    col *= inv_diag[i];
-                }
-            })
-        }
-    }
-
-    // NOTE: this actually computes the strain rates * density_i
-    fn compute_strain_rates(
-        &mut self,
-        dt: N,
-        fluid_fluid_contacts: &[ParticlesContacts<N>],
-        fluid_boundary_contacts: &[ParticlesContacts<N>],
-        fluids: &[Fluid<N>],
-        compute_error: bool,
-    ) -> N {
-        let boundaries_volumes = &self.boundaries_volumes;
-        let velocity_changes = &self.velocity_changes;
-        let min_neighbors_for_divergence_solve = self.min_neighbors_for_divergence_solve;
-        let mut max_error = N::zero();
-        let _2: N = na::convert(2.0f64);
-
-        for fluid_id in 0..fluids.len() {
-            let fluid_fluid_contacts = &fluid_fluid_contacts[fluid_id];
-            let fluid_boundary_contacts = &fluid_boundary_contacts[fluid_id];
-            let strain_rates_i = &mut self.strain_rates[fluid_id];
-            let boundary_volumes = &self.boundaries_volumes;
-            let fluid_i = &fluids[fluid_id];
-
-            let err = par_iter_mut!(strain_rates_i).enumerate().fold(
-                N::zero(),
-                |curr_err, (i, strain_rates_i)| {
-                    let mut boundary_rate = StrainRate::zeros();
-                    let mut fluid_rate = StrainRate::zeros();
-
-                    for c in fluid_fluid_contacts.particle_contacts(i) {
-                        let fluid_j = &fluids[c.j_model];
-                        let v_i = fluid_i.velocities[c.i] + velocity_changes[c.i_model][c.i];
-                        let v_j = fluid_j.velocities[c.j] + velocity_changes[c.j_model][c.j];
-                        let v_ji = v_j - v_i;
-                        let rate = compute_strain_rate(&c.gradient, &v_ji);
-
-                        fluid_rate += rate * (fluid_j.particle_mass(c.j) / _2);
-                    }
-
-                    //                    for c in fluid_boundary_contacts.particle_contacts(i) {
-                    //                        let v_i = fluid_i.velocities[c.i] + velocity_changes[c.i_model][c.i];
-                    //                        // FIXME: take the boundary velocity into account?
-                    //                        let v_ji = -v_i;
-                    //                        let rate = compute_strain_rate(&c.gradient, &v_ji);
-                    //
-                    //                        boundary_rate +=
-                    //                            rate * (boundary_volumes[c.j_model][c.j] * fluid_i.density0 / _2);
-                    //                    }
-
-                    if compute_error {
-                        strain_rates_i.error = fluid_rate + boundary_rate - strain_rates_i.target;
-                        curr_err + strain_rates_i.error.norm_squared()
-                    } else {
-                        strain_rates_i.target = fluid_rate * fluid_i.viscosity
-                            + boundary_rate * fluid_i.boundary_viscosity;
-                        N::zero()
-                    }
-                },
-            );
-
-            let nparts = fluids[fluid_id].num_particles();
-            if nparts != 0 {
-                max_error = max_error.max(err / na::convert(nparts as f64));
-            }
-        }
-
-        max_error
-    }
-
-    fn compute_velocity_changes_for_viscosity(
-        &mut self,
-        dt: N,
-        inv_dt: N,
-        fluid_fluid_contacts: &[ParticlesContacts<N>],
-        fluid_boundary_contacts: &[ParticlesContacts<N>],
-        fluids: &[Fluid<N>],
-        boundaries: &[Boundary<N>],
-    ) {
-        let alphas = &self.alphas;
-        let boundaries_volumes = &self.boundaries_volumes;
-        let strain_rates = &self.strain_rates;
-        let betas = &self.betas;
-        let _2: N = na::convert(2.0);
-
-        for (fluid_id, fluid1) in fluids.iter().enumerate() {
-            par_iter_mut!(self.velocity_changes[fluid_id])
-                .enumerate()
-                .for_each(|(i, velocity_change)| {
-                    let ui = betas[fluid_id][i] * strain_rates[fluid_id][i].error;
-
-                    for c in fluid_fluid_contacts[fluid_id].particle_contacts(i) {
-                        let fluid2 = &fluids[c.j_model];
-                        let uj = betas[c.j_model][c.j]
-                            * strain_rates[c.j_model][c.j].error
-                            * (fluid2.density0 / fluid1.density0);
-                        let gradient = compute_gradient_matrix(&c.gradient);
-
-                        // Compute velocity change.
-                        let coeff = (ui + uj) * (fluid1.particle_mass(c.i) / _2);
-                        *velocity_change += gradient.tr_mul(&coeff) * fluid1.particle_mass(c.i);
-                    }
-
-                    //                    for c in fluid_boundary_contacts[fluid_id].particle_contacts(i) {
-                    //                        let boundary2 = &boundaries[c.j_model];
-                    //
-                    //                        // Compute velocity change.
-                    //                        let gradient = compute_gradient_matrix(&c.gradient);
-                    //
-                    //                        // Compute velocity change.
-                    //                        let coeff =
-                    //                            ui * (boundaries_volumes[c.j_model][c.j] * fluid1.density0 / _2);
-                    //                        let delta = gradient.tr_mul(&coeff) * fluid1.particle_mass(c.i);
-                    //                        *velocity_change += delta;
-                    //
-                    //                        // Apply the force to the boundary too.
-                    //                        let particle_mass = fluid1.volumes[c.i] * fluid1.density0;
-                    //                        boundary2.apply_force(c.j, delta * (inv_dt * particle_mass));
-                    //                    }
-                })
-        }
-    }
-
-    fn viscosity_solve(
-        &mut self,
-        dt: N,
-        inv_dt: N,
-        kernel_radius: N,
-        contact_manager: &mut ContactManager<N>,
-        fluids: &mut [Fluid<N>],
-        boundaries: &[Boundary<N>],
-    ) {
-        let _ = self.compute_betas(
-            inv_dt,
-            &contact_manager.fluid_fluid_contacts,
-            &contact_manager.fluid_boundary_contacts,
-            fluids,
-        );
-        let _ = self.compute_strain_rates(
-            dt,
-            &contact_manager.fluid_fluid_contacts,
-            &contact_manager.fluid_boundary_contacts,
-            fluids,
-            false,
-        );
-
-        for i in 0..self.max_viscosity_iter {
-            let avg_err = self.compute_strain_rates(
-                dt,
-                &contact_manager.fluid_fluid_contacts,
-                &contact_manager.fluid_boundary_contacts,
-                fluids,
-                true,
-            );
-            println!(
-                "Average viscosity error: {}, break after niters: {}",
-                avg_err, i
-            );
-            if avg_err <= self.max_viscosity_error && i >= self.min_viscosity_iter {
-                break;
-            }
-
-            self.compute_velocity_changes_for_viscosity(
-                dt,
-                inv_dt,
-                &contact_manager.fluid_fluid_contacts,
-                &contact_manager.fluid_boundary_contacts,
-                fluids,
-                boundaries,
-            );
-        }
-    }
-
     /// Solves pressure and non-pressure force for the given fluids and boundaries.
     ///
     /// Both `self.init_with_fluids` and `self.init_with_boundaries` must be called before this
@@ -1018,6 +680,16 @@ where
             fluids,
         );
 
+        self.surface_tension.solve(
+            dt,
+            inv_dt,
+            kernel_radius,
+            contact_manager,
+            fluids,
+            &self.densities,
+            &mut self.velocity_changes,
+        );
+
         self.compute_alphas(
             inv_dt,
             &contact_manager.fluid_fluid_contacts,
@@ -1039,8 +711,6 @@ where
             .iter_mut()
             .for_each(|vs| vs.iter_mut().for_each(|v| v.fill(N::zero())));
 
-        //        self.nonpressure_solve(dt, inv_dt, contact_manager, fluids);
-
         self.pressure_solve(
             dt,
             inv_dt,
@@ -1050,13 +720,13 @@ where
             boundaries,
         );
 
-        self.viscosity_solve(
+        self.viscosity.solve(
             dt,
             inv_dt,
             kernel_radius,
             contact_manager,
             fluids,
-            boundaries,
+            &mut self.velocity_changes,
         );
 
         self.update_positions(dt, fluids);
