+++ src/solver/viscosity/mod.rs
@@ -1,0 +1,3 @@
+pub use self::dfsph_viscosity::DFSPHViscosity;
+
+mod dfsph_viscosity;
